requirements.txt
fair-esm
runpod
torch==2.2.0
numpy==1.26.4
gdown
lightning